demo/requirements.txt
django-intercoolerjs
